pyproject.toml
[project]
name = "home-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
